requirements.txt
sseclient
kafka-python